+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2.5.0 (????-??-??)
+## 2.5.0 (2020-04-02)
 
 - Add log messages for error conditions which send an error message back to the calling client to `Gos\Bundle\WebSocketBundle\Server\App\Dispatcher\RpcDispatcher`
 - Add new log messages and extra error checks to `Gos\Bundle\WebSocketBundle\Server\App\Dispatcher\TopicDispatcher`